% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{annotate_maf_classes}
\alias{annotate_maf_classes}
\title{Function to update an unannotated MAF with classification information for use in MetaboLights.}
\usage{
annotate_maf_classes(maf_input, maf_output)
}
\arguments{
\item{maf_input}{The unannotated MAF file}

\item{maf_output}{The filename of the resulting MAF}
}
\description{
This function updates an unannotated MAF with classification information. It should be run second on an unannotated MAF.
}
