% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{annotate_maf_compounds}
\alias{annotate_maf_compounds}
\title{Function to annotate a MAF for use in MetaboLights.}
\usage{
annotate_maf_compounds(
  maf_input,
  maf_output,
  polarity,
  xcms_id,
  pol_mode,
  smiles,
  names
)
}
\arguments{
\item{maf_input}{The input MAF file}

\item{maf_output}{The filename of the resulting MAF}

\item{polarity}{The polarity used ("neg" or "pos")}

\item{xcms_id}{The id of XCMS peak detection, usually the FTxxxxx name}

\item{pol_mode}{The polarization mode for any feature (e.g. rep("neg",nrow(ms1_def_neg))}

\item{smiles}{Vector containing the SMILES information}

\item{names}{The UPAC names of the annotated compounds}
}
\description{
This function annotates a MAF with structure identification. It should be run second on an unannotated MAF or third on a MAF with classification information.
}
