% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{export_maf}
\alias{export_maf}
\title{Function to export a peak table to MAF for use in MetaboLights.}
\usage{
export_maf(peak_list, maf_filename)
}
\arguments{
\item{peak_list}{The peak table with MS1 features in rows and information and samples in columns}

\item{maf_filename}{The filename of the MAF file}
}
\description{
This function exports a peak table to MAF without any annotations. It should be run first.
}
