% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gap_weighting}
\alias{gap_weighting}
\title{Function to perform gap weighting on continuous values.}
\usage{
gap_weighting(morph_char, states = 8)
}
\arguments{
\item{morph_char}{A vector with character values}

\item{states}{Number of states, defaults to 8}
}
\description{
This function converts continuous quantitative traits to distinct character states useful for calculating phylogenies.
Here, we apply the gap weighting algorithm by Thiele (1993), https://doi.org/10.1006/clad.1993.1020 .
}
\examples{
gap_weighting(marchantiales$char_list$thallus.width, states=8)
}
