% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{heatmap.selected_features}
\alias{heatmap.selected_features}
\title{Function to draw a heatmap of selected variables}
\usage{
heatmap.selected_features(
  feat_list,
  sel_feat,
  sel_names = NULL,
  sample_colors = NULL,
  filename,
  main,
  scale = "col",
  plot_width = 6,
  plot_height = 5,
  cex_col = 0.5,
  cex_row = 0.7
)
}
\arguments{
\item{feat_list}{the matrix containing the molecular information, with samples in rows and molecular features in columns}

\item{sel_feat}{vector of selected variables}

\item{sel_names}{vector of names of selected variables}

\item{sample_colors}{color vector for the factor}

\item{filename}{plot to external pdf, use NULL for internal plotting}

\item{main}{title of the plot}

\item{scale}{scaling of columns or rows, defaults to "col"}

\item{plot_width}{width of plot, defaults to 6}

\item{plot_height}{height of plot, defaults to 5}

\item{cex_col}{size for column, defaults to 0.5}

\item{cex_row}{size for row, defaults to 0.7}
}
\description{
Function to draw a heatmap of selected variables
}
