% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make.high_level.descriptor}
\alias{make.high_level.descriptor}
\title{Function to make abstracted high-level molecular descriptors from CDK names.}
\usage{
make.high_level.descriptor(descriptor_names)
}
\arguments{
\item{descriptor_names}{A vector containing the names of the molecular descriptors}
}
\description{
Function to make abstracted high-level molecular descriptors from CDK names.
}
