% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_classes_at_level}
\alias{make_classes_at_level}
\title{Function to extract information at a specific ontology level or below of classification information.}
\usage{
make_classes_at_level(level = 2, div_samples)
}
\arguments{
\item{level}{The level of the ontology to extract information to}

\item{div_samples}{Data frame with classes in rows and samples in columns.}
}
\description{
Function to extract information at a specific ontology level or below of classification information.
}
