% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{menhinick.diversity}
\alias{menhinick.diversity}
\title{Function to calculate the Menhinick diversity measure}
\usage{
menhinick.diversity(p)
}
\arguments{
\item{p}{vector with response variables of one sample}
}
\description{
Function to calculate the Menhinick diversity measure
}
\examples{
menhinick.diversity(p=c(4,8))
}
