% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{select_features_lasso}
\alias{select_features_lasso}
\title{Function for variables selection using the LASSO method}
\usage{
select_features_lasso(feat_matrix, sel_factor, family = "gaussian")
}
\arguments{
\item{feat_matrix}{the matrix containing the molecular information, with samples in rows and molecular features in columns}

\item{sel_factor}{the factor which is used to select the variables}

\item{family}{the distribution, always use "gaussian" here}
}
\description{
Function for variables selection using the LASSO method
}
\examples{
select_features_lasso(feat_matrix=tenerife$comp_list, sel_factor=tenerife$species)
}
