% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{select_features_pls}
\alias{select_features_pls}
\title{Function for variables selection using the Elastic Net method}
\usage{
select_features_pls(
  feat_matrix,
  sel_factor,
  sel_colors,
  components = 2,
  tune_length = 10,
  quantile_threshold = 0.95,
  plot_roc_filename = NULL
)
}
\arguments{
\item{feat_matrix}{the matrix containing the molecular information, with samples in rows and molecular features in columns}

\item{sel_factor}{the factor which is used to select the variables}

\item{sel_colors}{color vector for the factor}

\item{tune_length}{how often repeated cross validation is performed, defaults to 100}

\item{plot_roc_filename}{plot metrics to external pdf, defaults to NULL}

\item{confidence}{confidence level, defaults to 0.1}
}
\description{
Function for variables selection using the Elastic Net method
}
\examples{
select_features_pls(feat_matrix=tenerife$comp_list, sel_factor=tenerife$species, sel_colors=sapply(tenerife$species, function(x) { x <- RColorBrewer::brewer.pal(n=nlevels(tenerife$species), name="Set1")[which(x==levels(tenerife$species))] } ), components=5)
}
