% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{shannon.diversity}
\alias{shannon.diversity}
\title{Function to calculate the Shannon diversity measure}
\usage{
shannon.diversity(p)
}
\arguments{
\item{p}{vector with response variables of one sample}
}
\description{
This function is now obsolete, as it returns the same values than vegan::diversity(x, index="shannon"). Since we do not expect variables with negative intensity, we exclude negative values and take the natural logarithm.
}
\examples{
shannon.diversity(p=c(4,8))
}
