% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sunBurstPlotFromSubstanceClasses}
\alias{sunBurstPlotFromSubstanceClasses}
\title{Function to draw a sunburst plot}
\usage{
sunBurstPlotFromSubstanceClasses(
  classifierClasses,
  numberOfSpectra,
  colorStart = 0,
  colorAlpha = 0.6
)
}
\arguments{
\item{classifierClasses}{compound class names}

\item{numberOfSpectra}{frequency of classes}

\item{colorStart}{color start, defaults to 0.0}

\item{colorAlpha}{color alpha, defaults to 0.6}
}
\description{
Function to draw a sunburst plot
}
