% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tukey.test}
\alias{tukey.test}
\title{Function to perform the Tukey post-hoc HSD test on response and model terms}
\usage{
tukey.test(response, term)
}
\arguments{
\item{response}{vector with response variables}

\item{term}{vector with factorized terms}
}
\description{
Function to perform the Tukey post-hoc HSD test on response and model terms
}
