% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{select_features_random_forest}
\alias{select_features_random_forest}
\title{Function for variables selection using Random Forest}
\usage{
select_features_random_forest(
  feat_matrix,
  sel_factor,
  sel_colors,
  tune_length = 10,
  quantile_threshold = 0.95,
  keepx_min = 10,
  keepx_max = 0,
  plot_roc_filename = NULL
)
}
\arguments{
\item{feat_matrix}{the matrix containing the molecular information, with samples in rows and molecular features in columns}

\item{sel_factor}{the factor which is used to select the variables}

\item{sel_colors}{color vector for the factor}

\item{tune_length}{how often repeated cross validation is performed, defaults to 10}

\item{quantile_threshold}{confidence level, defaults to 0.95}

\item{keepx_min}{Minimum number of variables to keep for level (defaults to 10)}

\item{keepx_max}{Maximum number of variables to keep for level (defaults to 0, which means no limit)}

\item{plot_roc_filename}{plot metrics to external pdf, defaults to NULL}
}
\description{
Function for variables selection using Random Forest
}
