% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{shannon.diversity}
\alias{shannon.diversity}
\title{Function to calculate the Shannon diversity measure}
\usage{
shannon.diversity(p)
}
\arguments{
\item{p}{vector with response variables of one sample}
}
\description{
Function to calculate the Shannon diversity measure
}
\examples{
shannon.diversity(p=c(4,8))
}
